Login Implemented
Implemented a basic login and authorization system. It is currently
implementing test versions of the IUser, and IUserDbContext interfaces.
We will implement a database implementation of the IUserDbContext when I
have more information of the database structure.

Added the @authorize() decorator to provide authorization checking for
users.

Reformatted code to match the standards presented in the PEP-8 style
guide.